Refactor: Update namespaces for IFilter and SortBase
